Fix Deadline & Event task names
Add parseName method which combines all strings
after command to get task name

Update parseToDoCommand, parseDeadlineCommand, parseEventCommand to
use parseName method to get taskName

To improve abstraction and encapsulate the parsing logic so that in
future, if there is a new way to interpret commands (instead of '/')
then this method can be easily changed. Moreover, if there are new
types of tasks then they can use this method as well.